+++ Mailer/extensions/mailer-module/src/Api/v2/Handlers/Mailers/MailgunEventsHandler.php
@@ -83,6 +83,7 @@ class MailgunEventsHandler extends BaseHandler
             'reason' => $params['event-data']['reason'] ?? null,
             'severity' => $params['event-data']['severity'] ?? null,
             'client' => $params['event-data']['client-info'] ?? null,
+            'url' => $params['event-data']['url'] ?? null,
         ]), RedisDriver::PRIORITY_LOW);
 
         return new JsonApiResponse(200, ['status' => 'ok']);