+++ Mailer/extensions/mailer-module/src/Models/ContentGenerator/Replace/AnchorRtmReplace.php
@@ -18,7 +18,7 @@ class AnchorRtmReplace implements IReplace
         $this->allowedDomainManager = $allowedDomainManager;
     }
 
-    public function replace(string $content, GeneratorInput $generatorInput): string
+    public function replace(string $content, GeneratorInput $generatorInput, array $context = null): string
     {
         $matches = [];
         preg_match_all('/<a(.*?)href="([^"]*?)"(.*?)>/i', $content, $matches);