+++ Mailer/extensions/mailer-module/src/Models/ContentGenerator/Replace/IReplace.php
@@ -6,5 +6,11 @@ use Remp\MailerModule\Models\ContentGenerator\GeneratorInput;
 
 interface IReplace
 {
-    public function replace(string $content, GeneratorInput $generatorInput): string;
+    /**
+     * @param string $content
+     * @param GeneratorInput $generatorInput
+     * @param array|null $context random array with additional information for replacers
+     * @return string
+     */
+    public function replace(string $content, GeneratorInput $generatorInput, array $context = null): string;
 }